- Added face projection for 2D computations - Added error computation for exact solution - Added more test-cases - Improved code efficiency

+++ CleanTests.sh
@@ -12,11 +12,12 @@ function createLink() {
 
 HOME="$(pwd)"
 
-# declare -a BigCaseNames=("Neumann" "Dirichlet")
-declare -a BigCaseNames=("Dirichlet")
-declare -a TestCaseNames=("UnitSquare_Struct")
-declare -a SubTestCaseNames=("Coarse" "Medium" "Fine" "VeryFine")
-declare -a HowManySubCases=("4")
+declare -a BigCaseNames=("Neumann" "Dirichlet")
+# declare -a BigCaseNames=("Dirichlet")
+# declare -a TestCaseNames=("UnitSquare_Struct" "UnitCube_Struct")
+declare -a TestCaseNames=("UnitSquare_Struct" "UnitSquare_UnstructAligned" "UnitSquare_Unstruct" "UnitSquare_UnstructMixedAligned")
+declare -a SubTestCaseNames=("Coarse" "Medium" "Fine" "VeryFine" "Coarse" "Medium" "Fine" "VeryFine" "Coarse" "Medium" "Fine" "VeryFine" "Coarse" "Medium" "Fine" "VeryFine")
+declare -a HowManySubCases=("4" "4" "4" "4")
 
 
 
@@ -36,16 +37,21 @@ do
 
       SubCase=""
       cd ${HOME}/TestCases/${BigCaseName}/${TestCase}
-      echo "Cleaning test case ${BigCaseName}/${TestCase}/${SubCase}"
 
       if [ ${HowManySubCases[$iTestCase]} -eq "1" ]
       then
-         mv Parameters_ToUse.py ../dummy.py && rm -r * && mv ../dummy.py Parameters_ToUse.py
+        echo "Cleaning test case ${BigCaseName}/${TestCase}"
+        mv Parameters_ToUse.py ../dummy.py && rm -r * && mv ../dummy.py Parameters_ToUse.py
       else
         SubCase=${SubTestCaseNames[$globalSubCase]}
-        rm -rf $SubCase
-
+        echo "Cleaning test case ${BigCaseName}/${TestCase}/${SubCase}"
+        if [[ -d $SubCase ]]; then
+          rm -rf $SubCase
+        else
+          echo "Test Case ${BigCaseName}/${TestCase}/${SubCase} already cleaned"
+        fi
       fi
+      
 
 
       globalSubCase=$((globalSubCase+1))
